:sparkles: #37 별점 타입 Integer로 수정

+++ src/main/java/com/example/SEENEMA/post/view/domain/ViewPost.java
@@ -45,13 +45,13 @@ public class ViewPost {
     @Column(nullable = false)
     private String content; // 내용
     @Column
-    private Double viewScore;   // 시야 점수
+    private Integer viewScore;   // 시야 점수
     @Column
-    private Double seatScore;   // 좌석 점수
+    private Integer seatScore;   // 좌석 점수
     @Column
-    private Double lightScore;  // 조명 점수
+    private Integer lightScore;  // 조명 점수
     @Column
-    private Double soundScore;  // 음향 점수
+    private Integer soundScore;  // 음향 점수
 
     @OneToMany(cascade = CascadeType.ALL)
     private List<Image> image;
@@ -68,7 +68,7 @@ public class ViewPost {
 
     @Builder
     public ViewPost(User user, Theater theater, String play, String seat, String title, String content,
-                    Double viewScore, Double seatScore, Double lightScore, Double soundScore,List<Image> image, LocalDateTime createdAt) {
+                    Integer viewScore, Integer seatScore, Integer lightScore, Integer soundScore,List<Image> image, LocalDateTime createdAt) {
         this.user = user;
         this.theater = theater;
         this.play = play;
@@ -85,7 +85,7 @@ public class ViewPost {
     }
 
     public void updateViewPost(String play, String seat, String title, String content,
-                               Double viewScore, Double seatScore, Double lightScore, Double soundScore,List<Image> image) {
+                               Integer viewScore, Integer seatScore, Integer lightScore, Integer soundScore,List<Image> image) {
         this.play = play;
         this.seat = seat;
         this.title = title;
